Do not use "ingress-node-replication" with VXLAN
JTAC advises to not use "ingress-node-replication" when using BGP EVPN VXLAN. "multicast-mode ingress-replication" is enough to achieve the same effect. "ingress-node-replication" will replicate packets to VTEP not advertising the Type 3 route and also incurs a low-level overhead which may exhaust the chipset capacity (on a Broadcom platform for example). Please check with a QFX engineer if needed.